refactor!: Move config helpers from `model` to new `config-utils` module
Move the `ConfigurationResolver` and related helper functions to a new
`config-utils` module. This is preparation for being able to move the
`PackageConfigurationProvider` and `PackageCurationProvider` to the
respective API modules which is required for migrating them to the new
plugin API.
